requirements.txt
# ── Core ─────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# ── Async DB ─────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.20.0
asyncpg>=0.29.0       # PostgreSQL async driver (production)

# ── Redis (job queue & dedup cache) ─────────────────────────────────
redis[hiredis]>=5.0.0

# ── LLM providers (install the one you need) ────────────────────────
openai>=1.12.0
google-generativeai>=0.4.0

# ── Agent orchestration ─────────────────────────────────────────────
langgraph>=0.2.0
langchain-core>=0.3.0

# ── Frontend ─────────────────────────────────────────────────────────
streamlit>=1.31.0
requests>=2.31.0

# ── Dev / testing ────────────────────────────────────────────────────
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
